requirements.txt
tesseract
transformers
pytesseract
pillow
torch
gradio
streamlit